Pass the '/alerts' test
